+++ CHANGELOG.md
@@ -2,9 +2,13 @@
 
 ## Unreleased
 
-### Fix
+### Added
 
-* :warn: breaking change. All `fully_covered` returned values have changed from arrays
+* new feature `box_search`
+
+### Fixed
+
+* :warning: breaking change. All `fully_covered` returned values have changed from arrays
   of 0 and 255 to boolean arrays. This concerns: `box_search`, `cone_search`,
   `polygon_search`, `elliptical_cone_search`, and `elliptic-search`.
 * bilinear_interpolation now accepts longitudes and latitudes with nan values (will be a masked value in the output) [#22]